ya programo mejor q el profe de proyecto1 🔥🔥🔥🔥

+++ Project_9_Handout/Source/ModuleRender.cpp
@@ -93,7 +93,7 @@ bool ModuleRender::CleanUp()
 }
 
 // Blit to screen
-bool ModuleRender::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, float speed, bool useCamera)
+bool ModuleRender::Blit(SDL_Texture* texture, float x, float y, const SDL_Rect* section, float speed, bool useCamera)
 {
 	bool ret = true;
